+++ tests/testmain/test_main_classes.py
@@ -1,0 +1,6 @@
+
+model_props = {}
+handler = HandlerRCHRES(model_props)
+model = ModelRCHRES()
+handler.set_props(model, model_props)
+handler.init_nexits(model)
